infosec-backend/requirements.txt
Django>=4.2.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0
django-allauth>=0.54.0
dj-rest-auth>=5.0.0
djangorestframework-simplejwt>=5.2.0
python-decouple>=3.8
razorpay>=1.4.2
